% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introUtils.R
\name{assign.int.labels}
\alias{assign.int.labels}
\title{Convert string labels to integer labels}
\usage{
assign.int.labels(pop)
}
\arguments{
\item{pop}{a vector of integers or strings specifying which subpopulation the individuals belong to.}
}
\value{
The returned value is a vector of integers.
}
\description{
\code{assign.int.labels} returns a vector of integers.
}
