% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AX}
\alias{AX}
\title{A * x}
\usage{
AX(filename, x, nSNP, nIND)
}
\arguments{
\item{filename}{a character string.}

\item{x}{a numeric vector.}

\item{nSNP}{an integer.}

\item{nIND}{an integer.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{AX} computes the product A * x.
}
