% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bary_to_ancestry_cpp}
\alias{bary_to_ancestry_cpp}
\title{Barycentric coordinates to ancestry}
\usage{
bary_to_ancestry_cpp(scores, pop, popUnique, admixed)
}
\arguments{
\item{scores}{a numeric matrix.}

\item{pop}{a string vector.}

\item{popUnique}{a string vector.}

\item{admixed}{a character vector.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{centroids_to_simplex_cpp}
}
