% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed2matrix.R
\name{bed2matrix}
\alias{bed2matrix}
\title{Convert a bed to a matrix}
\usage{
bed2matrix(bedfile, n = NULL, p = NULL)
}
\arguments{
\item{bedfile}{Path to a bed file.}

\item{n}{Number of samples. Default reads it from corresponding fam file.}

\item{p}{Number of SNPs. Default reads it from corresponding bim file.}
}
\value{
An integer matrix.
}
\description{
Convert a bed to a matrix
}
\examples{
bedfile <- system.file("extdata", "geno3pops.bed", package = "pcadapt")
mat <- bed2matrix(bedfile)
dim(mat)
table(mat)
}
