% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedadapt.R
\name{BED_rsvd}
\alias{BED_rsvd}
\title{SVD for genotype matrices stored in .bed files with missing values}
\usage{
BED_rsvd(X, k)
}
\arguments{
\item{X}{an object of class bedadapt.}

\item{k}{an integer specifying the number of principal components to retain.}
}
\description{
\code{BED_rsvd}
}
