% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedadapt.R
\name{BED_rsvd_missing}
\alias{BED_rsvd_missing}
\title{SVD for genotype matrices stored in .bed files with missing values}
\usage{
BED_rsvd_missing(X, k, miss.ind, miss.snp)
}
\arguments{
\item{X}{an object of class bedadapt.}

\item{k}{an integer specifying the number of principal components to retain.}

\item{miss.ind}{a vector indicating the number of missing values for each 
individual.}

\item{miss.snp}{a vector indicating the number of missing values for each 
SNP.}
}
\description{
\code{BED_rsvd_missing}
}
