% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cart2bary_cpp}
\alias{cart2bary_cpp}
\title{Cartesian coordinates to barycentric coordinates}
\usage{
cart2bary_cpp(X, P)
}
\arguments{
\item{X}{a matrix.}

\item{P}{a matrix.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{cart2bary_cpp} returns the barycentric coordinates.
}
