% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{centroids_to_simplex_cpp}
\alias{centroids_to_simplex_cpp}
\title{Simplex}
\usage{
centroids_to_simplex_cpp(centroids, popUnique, admixed)
}
\arguments{
\item{centroids}{a numeric matrix.}

\item{popUnique}{a string vector.}

\item{admixed}{a character vector.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{centroids_to_simplex_cpp}
}
