% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_row}
\alias{check_row}
\title{Skip or discard}
\usage{
check_row(x, i)
}
\arguments{
\item{x}{a genotype matrix.}

\item{i}{an integer.}
}
\value{
The returned value is an integer.
}
\description{
\code{check_row} returns 0 for markers to be kept and 1 for markers to be discarded.
}
