% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{clumping}
\alias{clumping}
\title{LD clumping}
\usage{
clumping(G, chr.info, size = 100, thr = 0.2)
}
\arguments{
\item{G}{a genotype matrix.}

\item{chr.info}{a vector containing the chromosome information for each 
marker.}

\item{size}{an integer.}

\item{thr}{a numerical value.}
}
\value{
The returned value is a logical vector.
}
\description{
\code{clumping} is adapted from the snp_clumping function implemented in 
the bigsnpr package developed by Florian Prive.
}
