% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.R
\name{cmpt.cov.matrix}
\alias{cmpt.cov.matrix}
\title{Covariance for loaded genotype data}
\usage{
cmpt.cov.matrix(input, ploidy = 2)
}
\arguments{
\item{input}{a genotype matrix.}

\item{K}{an integer specifying the number of principal components to retain.}
}
\value{
The returned value is the covariance matrix.
}
\description{
Covariance for loaded genotype data
}

