% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_all_stat}
\alias{cmpt_all_stat}
\title{Introgression statistics}
\usage{
cmpt_all_stat(geno, V, sigma, window_size, direction, lab, ancstrl1, ancstrl2,
  adm, axis, map)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{V}{a loading matrix.}

\item{sigma}{a vector of singular values.}

\item{window_size}{an integer.}

\item{direction}{an integer.}

\item{lab}{a vector of integers.}

\item{ancstrl1}{an integer.}

\item{ancstrl2}{an integer.}

\item{adm}{an integer.}

\item{axis}{a numeric vector.}

\item{map}{a numeric vector containing the genetic positions.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{cmpt_all_stat} computes the statistics.
}
