% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_directional_stat}
\alias{cmpt_directional_stat}
\title{Directional statistics}
\usage{
cmpt_directional_stat(usc, uglob, lab, adm, ax)
}
\arguments{
\item{usc}{a matrix of rescaled local scores.}

\item{uglob}{a matrix of global scores.}

\item{lab}{a vector of integers.}

\item{adm}{an integer.}

\item{ax}{a numeric vector.}
}
\value{
The returned value is a numeric value.
}
\description{
\code{cmpt_directional_stat} computes the displacement of admixed 
individuals such that positive (resp. negative) displacement corresponds to 
a displacement towards ancestral population 2 (resp. 1).
}
