% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_global_pca}
\alias{cmpt_global_pca}
\title{Global Principal Component Analysis}
\usage{
cmpt_global_pca(geno, V, sigma)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{V}{a loading matrix.}

\item{sigma}{a vector of singular values.}
}
\value{
The returned value is a matrix of scores.
}
\description{
\code{cmpt_global_pca} computes the scores using all genetic markers.
}
