% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_local_pca}
\alias{cmpt_local_pca}
\title{Local Principal Component Analysis}
\usage{
cmpt_local_pca(geno, V, sigma, beg, end)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{V}{a loading matrix.}

\item{sigma}{a vector of singular values.}

\item{beg}{an integer specifying the first marker to be included.}

\item{end}{an integer specifying the first marker to be excluded.}
}
\value{
The returned value is a matrix of scores.
}
\description{
\code{cmpt_local_pca} computes the scores using a subset of genetic markers.
}
