% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_minor_af}
\alias{cmpt_minor_af}
\title{Minor allele frequencies}
\usage{
cmpt_minor_af(xmatrix, ploidy)
}
\arguments{
\item{xmatrix}{a genotype matrix.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{cmpt_minor_af} computes the minor allele frequencies.
}
