% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_new_win}
\alias{cmpt_new_win}
\title{\code{cmpt_new_win} computes the statistics.}
\usage{
cmpt_new_win(i, map, window_size)
}
\arguments{
\item{i}{an integer.}

\item{map}{a vector containing the genetic positions in Morgans.}

\item{window_size}{a numeric value specifying the window size en Morgans.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{cmpt_new_win} computes the statistics.
}
