% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_stat_introgr_bary}
\alias{cmpt_stat_introgr_bary}
\title{Introgression statistics}
\usage{
cmpt_stat_introgr_bary(geno, V, sigma, window_size, direction, lab, ancstrl1,
  ancstrl2, adm, axis, map, with_map, side)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{V}{a loading matrix.}

\item{sigma}{a vector of singular values.}

\item{window_size}{an integer.}

\item{direction}{an integer.}

\item{lab}{a vector of integers.}

\item{ancstrl1}{an integer.}

\item{ancstrl2}{an integer.}

\item{adm}{an integer.}

\item{axis}{a numeric vector.}

\item{map}{a numeric vector containing the genetic positions.}

\item{with_map}{an integer specifying whether the genetic positions have
been provided.}

\item{side}{an integer specifying whether the window should be aligned on 
the left, middle or right.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{cmpt_stat_introgr_bary} computes the statistics.
}
