% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_transformation}
\alias{cmpt_transformation}
\title{Match global centroids and local centroids}
\usage{
cmpt_transformation(uloc, uglob, lab, ancstrl1, ancstrl2, s, dloc, dglob, R)
}
\arguments{
\item{uloc}{a matrix of local scores.}

\item{uglob}{a matrix of global scores.}

\item{lab}{a vector of integers.}

\item{ancstrl1}{an integer.}

\item{ancstrl2}{an integer.}

\item{s}{a numeric vector.}

\item{dloc}{a numeric vector.}

\item{dglob}{a numeric vector.}

\item{R}{a numeric matrix.}
}
\description{
\code{cmpt_transformation} computes the local centroid, the global centroid and the scaling factor.
}
