% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_wilcoxon_stat}
\alias{cmpt_wilcoxon_stat}
\title{Wilcoxon statistics}
\usage{
cmpt_wilcoxon_stat(usc, uglob, direction, lab, adm, axis)
}
\arguments{
\item{usc}{a matrix of rescaled local scores.}

\item{uglob}{a matrix of global scores.}

\item{direction}{an integer.}

\item{lab}{a vector of integers.}

\item{adm}{an integer.}

\item{axis}{a numeric value.}
}
\value{
The returned value is a numeric value.
}
\description{
\code{cmpt_wilcoxon_stat} computes the Wilcoxon statistics.
}
