% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.R
\name{compute.covariance}
\alias{compute.covariance}
\title{Compute covariance matrix for large genotype data}
\usage{
compute.covariance(input.filename)
}
\arguments{
\item{input.filename}{a character string specifying the name of the file to be processed with \code{pcadapt}.}
}
\value{
The returned value is a square symmetric matrix.
}
\description{
Compute covariance matrix for large genotype data
}

