% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptClass.R
\name{compute.pval}
\alias{compute.pval}
\title{Principal Components p-values}
\usage{
compute.pval(st, K, method)
}
\arguments{
\item{st}{a chi-square statistic.}

\item{K}{an integer specifying the number of principal components to retain.}

\item{method}{a character string specifying the method to be used to compute
the p-values.}
}
\description{
\code{compute.pval} computes p-values for each genetic marker, depending on the
test statistics distribution.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}

