% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{computeStats}
\alias{computeStats}
\title{Test Statistics}
\usage{
computeStats(res, method, nSNP, K)
}
\arguments{
\item{res}{a list of quantities among which are the loadings.}

\item{method}{a character string that specifies the method used to compute the p-values.}

\item{nSNP}{an integer specifying the number of genetic markers.}

\item{K}{an integer specifying the number of principal components to retain.}
}
\description{
\code{computeStats} computes the test statistics for each genetic marker.
}
\examples{
## see ?pcadapt for examples
}
\keyword{internal}

