% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.R
\name{cov.std}
\alias{cov.std}
\title{Compute covariance matrix for loaded genotype data}
\usage{
cov.std(x, ploidy = 2)
}
\arguments{
\item{x}{a genotype matrix.}

\item{ploidy}{an integer.}
}
\value{
The returned value is a square symmetric matrix.
}
\description{
Compute covariance matrix for loaded genotype data
}

