% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolUtils.R
\name{cover.to.pool}
\alias{cover.to.pool}
\title{Simulate frequency matrix from genotype and coverage matrices}
\usage{
cover.to.pool(data, cover.matrix, pop, ploidy = 2)
}
\arguments{
\item{data}{a matrix with n rows and p columns where n is the number of individuals and p is the number of markers.}

\item{cover.matrix}{a matrix with n rows and p columns where n is the number of pools and is the number of markers.}

\item{pop}{a list of integers or strings specifying which subpopulation the individuals belong to.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}
}
\description{
\code{cover.to.pool} creates a matrix of frequency estimates, given a genotype matrix and a coverage matrix.
}

