% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{covRob_cpp}
\alias{covRob_cpp}
\title{Robust estimates for location and scatter}
\usage{
covRob_cpp(x)
}
\arguments{
\item{x}{a numeric matrix.}
}
\description{
\code{covRob_cpp} implements the Orthogonalized Gnanadesikan-Kettenring estimator
of Maronna and Zamar.
}
\examples{
## see also ?pcadapt for examples

}
