% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptClass.R
\name{create.pcadapt.matrix}
\alias{create.pcadapt.matrix}
\title{pcadapt objects}
\usage{
create.pcadapt.matrix(input, K, method, min.maf, ploidy)
}
\arguments{
\item{input}{a genotype matrix.}

\item{K}{an integer specifying the number of principal components to retain.}

\item{method}{a character string specifying the method to be used to compute
the p-values. Three statistics are currently available, \code{"mahalanobis"},
\code{"communality"}, and \code{"componentwise"}.}

\item{min.maf}{a value between \code{0} and \code{0.45} specifying the threshold
of minor allele frequencies above which p-values are computed.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}
}
\value{
The returned value is a list containing all the numerical quantities needed to create an object of class \code{pcadapt}.
}
\description{
\code{create.pcadapt.matrix} computes the numerical quantities needed to compute the test
statistics, and stores them in an object of class \code{pcadapt}.
}

