% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{display.local.pca}
\alias{display.local.pca}
\title{Local Ancestry}
\usage{
display.local.pca(geno, obj.svd, begin = 1, end = nrow(obj.svd$v), i = 1,
  j = 2, pop)
}
\arguments{
\item{geno}{a scaled genotype matrix..}

\item{obj.svd}{an object with \code{u}, \code{d} and \code{v}.}

\item{begin}{an integer.}

\item{end}{an integer.}

\item{i}{an integer.}

\item{j}{an integer.}

\item{pop}{a list of integers or strings specifying which subpopulation the 
individuals belong to.}
}
\description{
\code{loc.anc.plotting} plots the z-scores derived from the statistics 
computed with `scan.intro`.
}
\examples{
## see ?pcadapt for examples

}
