% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introUtils.R
\name{draw.pca}
\alias{draw.pca}
\title{Display local PCA}
\usage{
draw.pca(geno, V, sigma, uglob, beg, end, pop, i = 1, j = 2, ancstrl1,
  ancstrl2, adm)
}
\arguments{
\item{geno}{a genotype matrix.}

\item{V}{a loading matrix.}

\item{sigma}{a vector of singular values.}

\item{uglob}{a matrix of global scores.}

\item{beg}{an integer specifying the first marker to be included.}

\item{end}{an integer specifying the first marker to be excluded.}

\item{pop}{a list of integers.}

\item{i}{an integer indicating onto which principal component the individuals
are projected when the "scores" option is chosen. Default value is set to 
\code{1}.}

\item{j}{an integer indicating onto which principal component the individuals
are projected when the "scores" option is chosen. Default value is set to 
\code{2}.}

\item{ancstrl1}{an integer.}

\item{ancstrl2}{an integer.}

\item{adm}{an integer.}
}
\value{
The returned value is a list containing the test statistics and the 
associated p-values.
}
\description{
\code{draw.pca} displays both local and global scores.
}
