% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{fdrCalc}
\alias{fdrCalc}
\title{False Discovery Rate}
\usage{
fdrCalc(x, list_selection)
}
\arguments{
\item{x}{a vector of indices corresponding to the discoveries.}

\item{list_selection}{a vector of indices containing the list of all the SNPs under selection.}
}
\description{
\code{fdrCalc} computes the false discovery rate (FDR) for a given list of discoveries,
knowing the list of the Single Nucleotide Polymorphisms (SNPs) under selection.
}
\examples{
x <- c(1,4,5,8,9)
list_selection <- c(4,5,7,10,11)
fdrCalc(x,list_selection)
}

