% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{fdrControl}
\alias{fdrControl}
\title{False Discovery Rate Control Plotting}
\usage{
fdrControl(x, alpha_thresholds = seq(0.01, 0.5, length = 100))
}
\arguments{
\item{x}{an output from \code{outlier} containing the false discovery rates for each threshold that is specified in \code{alpha_thresholds}.}

\item{alpha_threholds}{a vector containing the control thresholds. Default value set to \code{seq(0.01,0.50,length=100)}.}
}
\description{
\code{fdrControl} plots the false discovery rate as a function of thresholds under which the FDR are expected to be.
}
\examples{
x <- NULL
x$fdr <- seq(0.0,0.4,length=100)
fdrControl(x)
}
\keyword{internal}

