% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{fstGeno}
\alias{fstGeno}
\title{Fst auxiliary function}
\usage{
fstGeno(Geno, POPs, nIND, nPOP, PopSizes)
}
\arguments{
\item{Geno}{a vector of \code{0,1} or \code{2} of length \code{nIND}.}

\item{POPs}{a list of integers specifying which subpopulation the individuals belong to.}

\item{nPOP}{an integer specifying the number of SNPs present in the dataset.}

\item{PopSizes}{a list of length \code{nPOP} specifying the number of individuals for each subpopulation.}
}
\description{
\code{fstGeno} returns the Fst associated with a particular SNP using Weir and Cokram's formula
}
\examples{
## see also ?fstCalc for examples

}
\keyword{internal}

