% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{get.output.name}
\alias{get.output.name}
\title{Replace the file extension with the .pcadapt extension}
\usage{
get.output.name(name)
}
\arguments{
\item{name}{a character string specifying the name of the file to be
converted.}
}
\description{
\code{get.output.name} returns a character string specifying the name of the output file.
}
\examples{
## see also ?pcadapt for examples

}

