% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{get.pc}
\alias{get.pc}
\title{Get the principal component the most associated with a genetic marker}
\usage{
get.pc(x, list)
}
\arguments{
\item{x}{an object of class `pcadapt`.}

\item{list}{a list of integers corresponding to the indices of the markers of interest.}
}
\description{
\code{get.pc} returns a data frame such that each row contains the index of
the genetic marker and the principal component the most correlated with it.
If several components have the same maximum squared z-score, the first one
is returned. A missing value is returned only when all component z-scores
are missing for a marker.
}
