% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolUtils.R
\name{get.pool.matrix}
\alias{get.pool.matrix}
\title{Convert genotypes to pooled samples}
\usage{
get.pool.matrix(data, pop, ploidy = 2)
}
\arguments{
\item{data}{a matrix with n rows and p columns where n is the number of individuals and p is the number of markers.}

\item{pop}{a list of integers or strings specifying which subpopulation the individuals belong to.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}
}
\description{
\code{get.pool.matrix} creates a pooled-sequenced data out of a genotype matrix,
given the labels of each individuals.
}

