% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{get.pop.names}
\alias{get.pop.names}
\title{Retrieve population names}
\usage{
get.pop.names(pop)
}
\arguments{
\item{pop}{a list of integers or strings specifying which population the
individuals belong to.}
}
\description{
\code{get.pop.names} retrieves the population names from the population file.
}
\examples{
## see also ?pcadapt for examples

}
\keyword{internal}
