% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{get.score.color}
\alias{get.score.color}
\title{Population colorization}
\usage{
get.score.color(pop)
}
\arguments{
\item{pop}{a list of integers or strings specifying which population the
individuals belong to.}
}
\description{
\code{get.score.color} allows the user to display individuals of the same
predefined population with the same color when using the option
\code{"scores"} in \code{pcadapt}.
}
\examples{
## see also ?pcadapt for examples

}
\keyword{internal}
