% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_pop_size}
\alias{get_pop_size}
\title{Get population size}
\usage{
get_pop_size(pop, popUnique)
}
\arguments{
\item{pop}{a string vector.}

\item{popUnique}{a string vector.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{get_pop_size}
}
