% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_rank}
\alias{get_rank}
\title{Wilcoxon rank}
\usage{
get_rank(v_temp)
}
\arguments{
\item{v_temp}{a numeric vector.}
}
\value{
The returned value is a vector of integers.
}
\description{
\code{get_rank} returns the ordering index.
}
