% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_size_file}
\alias{get_size_file}
\title{File size}
\usage{
get_size_file(path)
}
\arguments{
\item{path}{a character string specifying the name of the file to be processed with \code{pcadapt}.}
}
\value{
The returned value is a numeric vector of length 2.
}
\description{
\code{get_size_file} returns the number of genetic markers and the number of individuals present in the data.
}

