% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_window}
\alias{get_window}
\title{\code{get_window}}
\usage{
get_window(i, map, window_size)
}
\arguments{
\item{i}{an integer.}

\item{map}{a vector containing the genetic positions in Morgans.}

\item{window_size}{a numeric value specifying the window size en Morgans.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{get_window}
}
