% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.R
\name{geteigen}
\alias{geteigen}
\title{SVD for large genotype matrices}
\usage{
geteigen(input.filename, K)
}
\arguments{
\item{input.filename}{a character string specifying the name of the file to be processed with \code{pcadapt}.}
}
\value{
The returned value is an object containing eigenvectors and eigenvalues.
}
\description{
SVD for large genotype matrices
}

