% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getPopColors}
\alias{getPopColors}
\title{Subpopulation colorization}
\usage{
getPopColors(pop)
}
\arguments{
\item{pop}{a list of integers or strings specifying which subpopulation the individuals belong to.}
}
\description{
\code{getPopColors} allows the user to display individuals of the same pre-defined subpopulation with the same color when using the option 
\code{"scores"} in \code{pcadapt}.
}
\examples{
## see also ?pcadapt for examples

}
\keyword{internal}

