% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaUtils.R
\name{getsize}
\alias{getsize}
\title{Return the number of rows and the the number of columns of large genotype matrices}
\usage{
getsize(input.filename)
}
\arguments{
\item{input.filename}{a character string specifying the name of the file to be processed with \code{pcadapt}.}
}
\value{
The returned value is a list of two integers.
}
\description{
Return the number of rows and the the number of columns of large genotype matrices
}

