% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{hist_plot}
\alias{hist_plot}
\title{Neutral Distribution Estimation}
\usage{
hist_plot(x, K)
}
\arguments{
\item{x}{an output from \code{outlier} containing the chi-squared statistics.}

\item{K}{an integer indicating which principal component the histogram will 
be associated with.}
}
\description{
\code{hist_plot} plots the histogram of the chi-squared statistics, as well 
as the estimated null distribution.
}
\keyword{internal}
