% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introUtils.R
\name{impute.pcadapt}
\alias{impute.pcadapt}
\title{Genotype matrix imputation}
\usage{
impute.pcadapt(input, pop, skip.return = FALSE)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of the file to be imputed.}

\item{pop}{a vector of integers or strings specifying which subpopulation the individuals belong to.}

\item{skip.return}{a logical value specifying whether the list of markers to be skipped should be returned or not.}
}
\value{
The returned value is a list containing the test statistics and the associated p-values.
}
\description{
\code{impute.pcadapt} imputes values based on medians.
}
