% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_geno}
\alias{impute_geno}
\title{Genotype matrix imputation}
\usage{
impute_geno(x)
}
\arguments{
\item{x}{a genotype matrix.}
}
\value{
The returned value is a list containing the imputed genotype matrix and a vector indicating which markers 
have to be discarded.
}
\description{
\code{impute_geno} imputes values based on the median per row.
}
