% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{impute_geno_pop}
\alias{impute_geno_pop}
\title{Genotype matrix imputation}
\usage{
impute_geno_pop(x, lab, pop)
}
\arguments{
\item{x}{a genotype matrix.}

\item{lab}{a vector of integers.}

\item{pop}{a vector of integers.}
}
\value{
The returned value is a list containing the imputed genotype matrix and a vector indicating which markers 
have to be discarded.
}
\description{
\code{impute_geno_pop} imputes values based on the median per row and per population.
}
