% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{kurtosis}
\alias{kurtosis}
\title{Kurtosis}
\usage{
kurtosis(x, prop = 0)
}
\arguments{
\item{x}{a numeric vector.}

\item{prop}{a value between \code{0} and \code{1} specifying the proportion of the distribution that is removed when computing the kurtosis.
Default value set to \code{0}.}
}
\description{
\code{kurtosis} computes the kurtosis of a given distribution \code{x}.
}
\examples{
x <- rnorm(10000)
kurtosis(x, prop = 0.2)
}

