% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lfmm2pcadapt}
\alias{lfmm2pcadapt}
\title{Convert lfmm files}
\usage{
lfmm2pcadapt(input, output)
}
\arguments{
\item{input}{a character string specifying the name of the file to be converted.}

\item{output}{a character string specifying the name of the output file.}
}
\description{
\code{lfmm2pcadapt} converts \code{lfmm} files to the format \code{pcadapt}.
}
\examples{
## see also ?pcadapt for examples

}
\keyword{internal}
