% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{loc.anc.plotting}
\alias{loc.anc.plotting}
\title{Local Ancestry}
\usage{
loc.anc.plotting(x, by.step = by.step, hline = hline)
}
\arguments{
\item{x}{an output from \code{outlier} containing the chi-squared statistics.}

\item{by.step}{an integer specifying the density}
}
\description{
\code{loc.anc.plotting} plots the z-scores derived from the statistics computed with `scan.intro`.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}
