% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introUtils.R
\name{logit.stat}
\alias{logit.stat}
\title{Introgression}
\usage{
logit.stat(input, pop, ancstrl.1, ancstrl.2, admxd, window.size = 1000)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of 
the file to be imputed.}

\item{pop}{a vector of integers or strings specifying which subpopulation the
individuals belong to.}

\item{ancstrl.1}{a string specifying the label of the ancestral population 
genetically closer to the hybrid population.}

\item{ancstrl.2}{a string specifying the label of the ancestral population 
genetically further from the hybrid population.}

\item{admxd}{a string specifying the label of the hybrid population.}

\item{window.size}{an integer specifying the window size.}
}
\value{
The returned value is a list containing the test statistics.
}
\description{
\code{logit.stat} computes statistics to detect excesses of local ancestry 
based on a logistic regression approach. The method currently supports only
two ancestral populations.
}
