% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit_fun}
\alias{logit_fun}
\title{Fast computation of rolling products.}
\usage{
logit_fun(x)
}
\arguments{
\item{x}{a genotype matrix.}
}
\value{
The returned value is a numerical vector.
}
\description{
\code{test}
}
