% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrUtils.R
\name{lrfunc}
\alias{lrfunc}
\title{Regression}
\usage{
lrfunc(matrix, filename, nind, nloci, K, ploidy, min.maf)
}
\arguments{
\item{filename}{a character string.}

\item{nind}{an integer.}

\item{nloci}{an integer.}

\item{K}{an integer.}

\item{ploidy}{an integer.}

\item{min.maf}{a real number.}

\item{mat}{a matrix.}
}
\description{
Regression
}

