% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lrfunc_cpp}
\alias{lrfunc_cpp}
\title{Compute graph matrix with the heat kernel function.}
\usage{
lrfunc_cpp(scores, Geno, K, nIND, nSNP, ploidy, min_maf)
}
\description{
Compute graph matrix with the heat kernel function.
}

