% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lrfunc_matrix}
\alias{lrfunc_matrix}
\title{Linear regression.}
\usage{
lrfunc_matrix(Geno, scores, nIND, nSNP, K, ploidy, min_maf)
}
\arguments{
\item{Geno}{a genotype matrix.}

\item{scores}{a matrix containing the scores.}

\item{nIND}{an integer specifying the number of individuals present in the data.}

\item{nSNP}{an integer specifying the number of genetic markers present in the data.}

\item{K}{an integer specifying the number of principal components to retain.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}

\item{min_maf}{a value between \code{0} and \code{0.45} specifying the threshold of minor allele frequencies above which p-values are computed.}
}
\value{
The returned value is a Rcpp::List containing the multiple linear regression z-scores, the minor allele frequencies and the number of missing values for each genetic marker.
}
\description{
\code{lrfunc_matrix} performs the multiple linear regression of the genotype matrix on the scores.
}

