% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{manhattan_plot}
\alias{manhattan_plot}
\title{Manhattan Plot}
\usage{
manhattan_plot(x, chr.info, snp.info, plt.pkg = "ggplot", K = 1)
}
\arguments{
\item{x}{an object of class "pcadapt" generated with \code{pcadapt} 
containing the p-values of interest.}

\item{chr.info}{a list containing the chromosome information for each marker.}

\item{snp.info}{a list containing the names of all genetic markers present in
the input.}

\item{plt.pkg}{a character string specifying the package to be used to 
display the graphical outputs. Use \code{"plotly"} for interactive plots, or 
\code{"ggplot"} for static plots.}

\item{K}{an integer specifying which principal component to display when 
\code{method="componentwise"}.}
}
\description{
\code{manhattan_plot} displays a Manhattan plot which represents the p-values 
for each SNP for a particular test statistic.
}
\keyword{internal}
