% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Manhattan Plot}
\usage{
manhattanPlot(x, K, alpha)
}
\arguments{
\item{x}{an object of class "outlier" generated with \code{outlier} containing the p-values of interest.}

\item{alpha}{a real number between \code{0} et \code{1}.}
}
\description{
\code{manhattanPlot} displays a Manhattan plot which represents the p-values for each SNP for a particular principal component (or set of
principal components in the communality test case).
}
\examples{
## see ?pcadapt for examples
}
\keyword{internal}

