% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedadapt.R
\name{matrix_rsvd}
\alias{matrix_rsvd}
\title{SVD for genotype matrices}
\usage{
matrix_rsvd(X, k, min.maf)
}
\arguments{
\item{X}{an object of class bedadapt.}

\item{k}{an integer specifying the number of principal components to retain.}

\item{min.maf}{a value between \code{0} and \code{0.45} specifying the 
threshold of minor allele frequencies above which p-values are computed.}
}
\description{
\code{matrix_rsvd}
}
