% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_per_pop}
\alias{median_per_pop}
\title{Median}
\usage{
median_per_pop(x, lab, pop, i)
}
\arguments{
\item{x}{a genotype matrix.}

\item{lab}{a vector of integers.}

\item{pop}{a vector of integers.}

\item{i}{an integer.}
}
\value{
The returned value is a real-valued vector.
}
\description{
\code{median_per_pop} computes the median of each population for a specific row.
}
