% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_row_i}
\alias{median_row_i}
\title{Median}
\usage{
median_row_i(x, i)
}
\arguments{
\item{x}{a genotype matrix.}

\item{i}{an integer.}
}
\value{
The returned value is a real number.
}
\description{
\code{median_row_i} computes the median of a specific row.
}
