% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{neutralDistribution}
\alias{neutralDistribution}
\title{Neutral Distribution Estimation}
\usage{
neutralDistribution(x, K, n_breaks = 100)
}
\arguments{
\item{K}{an integer indicating which principal component the histogram will be associated with.}

\item{n_breaks}{an integer indicating the number of bars the histogram will be composed of. Default value set to \code{100}.}

\item{data}{a numeric matrix which provides the data for the principal components analysis (PCA).}

\item{sigma}{a numeric vector containing the estimated standard deviations for each principal component.}
}
\description{
\code{neutralDistribution} plots the histogram of the loadings associated with the \code{K}-th principal component, as well
as the normal distribution with the estimated standard deviation.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}

