% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pca_rotation}
\alias{pca_rotation}
\title{\code{pca_rotation}}
\usage{
pca_rotation(a, b)
}
\arguments{
\item{a}{a numeric matrix.}

\item{b}{a numeric matrix.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{pca_rotation}
}
