\name{pcadapt-package}
\alias{pcadapt-package}
\docType{package}
\title{
	Principal Components Analysis for Outliers Detection.
}
\description{
This package has been developped to provide statistical tools for outliers detection based on Principal Components Analysis.
}
\details{
\tabular{ll}{
Package: \tab pcadapt\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-07-16\cr
License: \tab (>= 2)\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Keurcien Luu, Michael Blum

Maintainer: Keurcien Luu <keurcien.luu@imag.fr>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
}
\examples{
x<-NULL
}
