% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{pcadapt_verbose}
\alias{pcadapt_verbose}
\title{Summary}
\usage{
pcadapt_verbose(input, output, nIND, nSNP)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of the file to be converted.}

\item{output}{a character string specifying the name of the output file.}

\item{nIND}{an integer specifying the number of individuals present in the data.}

\item{nSNP}{an integer specifying the number of genetic markers present in the data.}
}
\description{
\code{pcadapt_verbose} prints out a summary of the file conversion.
}
\examples{
## see also ?pcadapt for examples

}

