% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.outlier}
\alias{plot.outlier}
\title{Outlier visualization tool}
\usage{
\method{plot}{outlier}(x, ..., option = NULL)
}
\arguments{
\item{x}{an object of class "outlier" generated with \code{outlier} containing the p-values of interest.}

\item{...}{\dots}

\item{option}{a character string specifying the figures to be displayed. If \code{NULL} (the default), all available plots are printed.}
}
\description{
\code{plot.outlier} is a method designed for objects of class \code{outlier}. For the moment,
the only option available is "fdr", which displays the observed false discovery rate as a function
of the expected false discovery rate.
}
\examples{
## see ?outlier for examples
}

