% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{print_convert}
\alias{print_convert}
\title{Summary}
\usage{
print_convert(input, output, M, N, pool)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of the file to be converted.}

\item{output}{a character string specifying the name of the output file.}

\item{M}{an integer specifying the number of genetic markers present in the data.}

\item{N}{an integer specifying the number of individuals present in the data.}

\item{pool}{an integer specifying the type of data. `0` for genotype data, `1` for pooled data.}
}
\description{
\code{print_convert} prints out a summary of the file conversion.
}
\keyword{internal}
