% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{pval}
\alias{pval}
\title{Principal Components p-values}
\usage{
pval(x, communality_test = FALSE)
}
\arguments{
\item{x}{an output of \code{pcadapt} containing the loadgins and the estimated standard deviations.}

\item{communality_test}{a logical value indicating whether a communality test should be performed. Default value set to \code{FALSE}.}
}
\description{
\code{pval} computes p-values for each principal component or for a set of principal components in the meta-analysis case.
}
\examples{
x <- replicate(1000, rnorm(100))
x <- floor(abs(x))
y <- pcadapt(x,K=3)
pval(y)
}
\keyword{internal}

