% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{pvalqq.plotting}
\alias{pvalqq.plotting}
\title{p-values Q-Q Plot}
\usage{
pvalqq.plotting(x, K, threshold)
}
\arguments{
\item{x}{an output from \code{outlier} containing the p-values of interest.}

\item{K}{an integer specifying which principal component to display when \code{method="componentwise"}.}

\item{threshold}{a real number between \code{0} et \code{1}.}
}
\description{
\code{pvalqq.plotting} plots a Q-Q plot of the p-values computed.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}

