% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{pvalqqPlot}
\alias{pvalqqPlot}
\title{p-values Q-Q Plot}
\usage{
pvalqqPlot(x, K, threshold = NULL)
}
\arguments{
\item{x}{an output from \code{outlier} containing the p-values of interest.}

\item{alpha}{a real number between \code{0} et \code{1}.}
}
\description{
\code{pvalqqPlot} plots a Q-Q plot of the p-values computed ).
}
\examples{
## see ?pcadapt for examples
}
\keyword{internal}

