% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{qq_plot}
\alias{qq_plot}
\title{p-values Q-Q Plot}
\usage{
qq_plot(x, K = 1)
}
\arguments{
\item{x}{an output from \code{outlier} containing the p-values of interest.}

\item{K}{an integer specifying which principal component to display when \code{method="componentwise"}.}
}
\description{
\code{qq_plot} plots a Q-Q plot of the p-values computed.
}
\keyword{internal}
