% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdadapt.R
\name{rdadapt}
\alias{rdadapt}
\title{Redundant Analysis for outlier detection}
\usage{
rdadapt(input, env)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of 
the file to be processed with \code{pcadapt}.}

\item{env}{a matrix of environmental or phenotypic variables.}
}
\value{
The returned value \code{x} is an object of class \code{pcadapt}.
}
\description{
\code{rdadapt}
}
