% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read.genotype}
\alias{read.genotype}
\title{Read files for pcadapt}
\usage{
read.genotype(file, option = NULL)
}
\arguments{
\item{file}{the name of the file without the extention which the data are to be read from.}

\item{option}{a character string that specifies which type of data the user is willing to read.}
}
\description{
\code{read.genotype} returns the genotype matrix for different types of files.
}
\keyword{internal}

