% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{read4pcadapt}
\alias{read4pcadapt}
\title{Multi-purpose read function}
\usage{
read4pcadapt(x = NULL, option = "genotype")
}
\arguments{
\item{x}{can be a file path or a file name, depending on the chosen option.}

\item{option}{a character string specifying the type of data going to be read.}
}
\description{
\code{read4pcadatp} allows the user to display individuals of the same pre-defined subpopulation with the same color when using the option
\code{"scores"} in \code{pcadapt}. Creates a global variable storing the lists of indices for each subpopulation.
}
\examples{
## see also ?pcadapt for examples
}

