% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rescale_local_pca}
\alias{rescale_local_pca}
\title{Rescale local scores}
\usage{
rescale_local_pca(u, s, dep_loc, dep_glob, R)
}
\arguments{
\item{u}{a matrix of scores.}

\item{s}{a numeric vector.}

\item{dep_loc}{a numeric vector.}

\item{dep_glob}{a numeric vector.}

\item{R}{a numeric matrix.}
}
\value{
The returned value is a list.
}
\description{
\code{rescale_local_pca} returns the rescaled local scores.
}
