% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sandbox.R
\name{residuals_to_stat}
\alias{residuals_to_stat}
\title{Residuals-based statistics}
\usage{
residuals_to_stat(geno, obj.svd, K = 1, pop, admixed, window.size = 100)
}
\arguments{
\item{obj.svd}{an object with \code{u}, \code{d} and \code{v}.}

\item{K}{an integer.}

\item{pop}{a list of integers or strings specifying which subpopulation the 
individuals belong to.}

\item{admixed}{a string specifying the label of the hybrid population.}

\item{window.size}{an integer.}

\item{input}{a scaled genotype matrix.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{residuals_to_stat}
}
