% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{roll_prod}
\alias{roll_prod}
\title{Fast computation of rolling products.}
\usage{
roll_prod(x, regcoeff, window_size)
}
\arguments{
\item{x}{a genotype matrix.}

\item{regcoeff}{a numerical vector.}

\item{window_size}{an integer.}
}
\value{
The returned value is a numerical vector.
}
\description{
\code{roll_prod}
}
