% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rsvd_cpp}
\alias{rsvd_cpp}
\title{Covariance for loaded genotype data}
\usage{
rsvd_cpp(A, k)
}
\arguments{
\item{A}{a numeric matrix.}

\item{k}{an integer.}
}
\value{
The returned value is a Rcpp::List.
}
\description{
\code{rsvd_cpp} computes the randomized SVD.
}
