% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poolUtils.R
\name{sample.geno}
\alias{sample.geno}
\title{Sample genotype matrix from pooled samples}
\usage{
sample.geno(pool.matrix = NULL, ploidy = 2, cover.matrix = NULL,
  pop.sizes = NULL, method = "per.pop")
}
\arguments{
\item{pool.matrix}{a matrix with n rows and p columns where n is the number of pools and is the number of markers.}

\item{ploidy}{an integer specifying the ploidy.}

\item{cover.matrix}{a matrix with n rows and p columns where n is the number of pools and is the number of markers.}

\item{pop.sizes}{a list specifying the number of individuals for each pool.}

\item{method}{a character string indicating the method used for sampling.}
}
\description{
\code{sample.geno} samples a genotype matrix from pooled samples.
}
\examples{
## see also ?pcadapt for examples

}

