% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_geno_cpp}
\alias{sample_geno_cpp}
\title{Sample genotype matrix from pooled samples}
\usage{
sample_geno_cpp(freq, ploidy, sample_size)
}
\description{
Sample genotype matrix from pooled samples
}

