% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_geno_file}
\alias{sample_geno_file}
\title{Sample genotype matrix from pooled samples}
\usage{
sample_geno_file(input, output, ploidy, sample_size)
}
\arguments{
\item{input}{a character string specifying the name of the file containing the allele frequencies.}

\item{output}{a character string specifying the name of the output file.}

\item{ploidy}{an integer specifying the ploidy of the sampled individuals.}

\item{sample_size}{a vector specifying the number of individuals to be sampled for each pool.}
}
\value{
The returned value is a numeric vector of length 2.
}
\description{
\code{sample_geno_file} sample genotypes based on observed allelic frequencies.
}

