% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale_geno}
\alias{scale_geno}
\title{Scale genotype matrices}
\usage{
scale_geno(xmatrix, ploidy, maf, keep_or_not)
}
\arguments{
\item{xmatrix}{a genotype matrix.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}

\item{maf}{a vector of minor allele frequencies.}

\item{keep_or_not}{a vector of integers.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{scale_geno} scales the genotype matrix.
}
