% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{score.plotly}
\alias{score.plotly}
\title{Principal Components Analysis Scores Plot}
\usage{
score.plotly(x, i = 1, j = 2, pop, gg.col, ind)
}
\arguments{
\item{x}{an output from \code{pcadapt} containing the scores.}

\item{i}{an integer indicating onto which principal component the individuals
 are projected when the "scores" option is chosen.
Default value is set to \code{1}.}

\item{j}{an integer indicating onto which principal component the individuals 
are projected when the "scores" option is chosen.
Default value is set to \code{2}.}

\item{pop}{a list of integers or strings specifying which subpopulation the 
individuals belong to.}

\item{ind}{a list of integers or strings specifying the names of the 
individuals.}
}
\description{
\code{"score.plotly"} plots the projection of the individuals onto the 
first two principal components.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}
