% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{score.plotting}
\alias{score.plotting}
\title{Principal Components Analysis Scores Plot}
\usage{
score.plotting(x, i = 1, j = 2, pop)
}
\arguments{
\item{x}{an output from \code{pcadapt} containing the scores.}

\item{i}{an integer indicating onto which principal component the individuals are projected when the "scores" option is chosen.
Default value is set to \code{1}.}

\item{j}{an integer indicating onto which principal component the individuals are projected when the "scores" option is chosen.
Default value is set to \code{2}.}

\item{pop}{a list of integers or strings specifying which subpopulation the individuals belong to.}
}
\description{
\code{"score.plotting"} plots the projection of the individuals onto the first two principal components.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}

