% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{score_plot}
\alias{score_plot}
\title{Principal Components Analysis Scores Plot}
\usage{
score_plot(x, i = 1, j = 2, pop, col, plt.pkg = "ggplot")
}
\arguments{
\item{x}{an output from \code{pcadapt} containing the scores.}

\item{i}{an integer indicating onto which principal component the individuals
 are projected when the "scores" option is chosen.
Default value is set to \code{1}.}

\item{j}{an integer indicating onto which principal component the individuals 
are projected when the "scores" option is chosen.
Default value is set to \code{2}.}

\item{pop}{a list of integers or strings specifying which subpopulation the 
individuals belong to.}

\item{plt.pkg}{a character string specifying the package to be used to 
display the graphical outputs. Use \code{"plotly"} for interactive plots, or 
\code{"ggplot"} for static plots.}
}
\description{
\code{"score_plot"} plots the projection of the individuals onto the 
first two principal components.
}
\keyword{internal}
