% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scores_centroids_cpp}
\alias{scores_centroids_cpp}
\title{Centroids}
\usage{
scores_centroids_cpp(scores, pop, popUnique)
}
\arguments{
\item{scores}{a numeric matrix.}

\item{pop}{a string vector.}

\item{popUnique}{a string vector.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{scores_centroids_cpp} returns the centroids of each population.
}
