% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{scree.plotting}
\alias{scree.plotting}
\title{Principal Components Analysis Scree Plot}
\usage{
scree.plotting(x, K)
}
\arguments{
\item{x}{an output from \code{pcadapt} containing the singular values.}

\item{K}{an integer specifying the number of components to take into account in the scree plot.}
}
\description{
\code{scree.plotting} plots the scee plot associated with the principal components analysis performed on the dataset.
NB : \code{pcadapt} has to be run on the dataset in order to get an output readable by \code{plot.screePlot}
}
\examples{
## see ?fastpcadapt for examples

}
\keyword{internal}

