% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtils.R
\name{scree_plot}
\alias{scree_plot}
\title{Principal Components Analysis Scree Plot}
\usage{
scree_plot(x, K = NULL)
}
\arguments{
\item{x}{an output from \code{pcadapt} containing the singular values.}

\item{K}{an integer specifying the number of components to plot.}
}
\description{
\code{scree_plot} plots the scree plot associated with the principal components 
analysis performed on the dataset. NB : \code{pcadapt} has to be run on the
dataset in order to get an output readable by \code{plot.screePlot}
}
\keyword{internal}
