% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{screePlot}
\alias{screePlot}
\title{Principal Components Analysis Scree Plot}
\usage{
screePlot(x, num_pc = NULL)
}
\arguments{
\item{x}{an output from \code{pcadapt} containing the singular values.}

\item{num_pc}{an integer specifying the number of components to take into account in the scree plot. \code{num_axis} should be lower than \code{K}.
By default, \code{num_axis = K}.}
}
\description{
\code{screePlot} plots the scee plot associated with the principal components analysis performed on the dataset.
NB : \code{pcadapt} has to be run on the dataset in order to get an output readable by \code{plot.screePlot}
}
\examples{
## see ?pcadapt for examples
}
\keyword{internal}

