% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{sdevEstimation}
\alias{sdevEstimation}
\title{Kurtosis-based Standard Deviation Estimation}
\usage{
sdevEstimation(x, method = "kurtosis.null")
}
\arguments{
\item{x}{a numeric vector whose distribution is supposed to be normal.}

\item{method}{a character string that specifies which method is used to estimate the standard deviations. Default value set to \code{kurtosis.null}.
An alternative option is \code{kurtosis.truncated}.}
}
\description{
\code{sdevEstimation} returns an estimate of the standard deviation of a set of values \code{x} which is supposed to follow a normal distribution.
}
\examples{
x <- rnorm(10000,sd=0.9)
sdevEstimation(x)
}

