% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{slidingWindows}
\alias{slidingWindows}
\title{Simplex}
\usage{
slidingWindows(sgeno, d, v, pop, popUnique, admixed, window_size, map)
}
\arguments{
\item{sgeno}{a scaled genotype matrix.}

\item{d}{a numeric vector.}

\item{v}{a numeric matrix.}

\item{pop}{a string vector.}

\item{popUnique}{a string vector.}

\item{admixed}{a character vector.}

\item{window_size}{a numeric value.}

\item{map}{a numeric vector.}
}
\value{
The returned value is a numeric matrix.
}
\description{
\code{centroids_to_simplex_cpp}
}
