% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tAA_cpp}
\alias{tAA_cpp}
\title{Product of a matrix with its transpose}
\usage{
tAA_cpp(x, nrow, ncol)
}
\arguments{
\item{x}{a matrix.}

\item{nrow}{an integer specifying the number of rows of \code{x}.}

\item{ncol}{an integer specifying the number of columns of \code{x}.}
}
\value{
The returned value is the result of the product of \code{x} with its transpose.
}
\description{
\code{tAA_cpp} computes the product of a real-valued matrix with its transpose.
}

