% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tAX}
\alias{tAX}
\title{A.t * x}
\usage{
tAX(filename, x, nSNP, nIND)
}
\arguments{
\item{filename}{a character string.}

\item{x}{a numeric vector.}

\item{nSNP}{an integer.}

\item{nIND}{an integer.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{tAX} computes the product A.t * x.
}
