% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{vcf2pcadapt}
\alias{vcf2pcadapt}
\title{vcfR-based converter}
\usage{
vcf2pcadapt(geno, output.file = "tmp.pcadapt", allele.sep = "/",
  blocksize = 10000, console.count = 10000)
}
\arguments{
\item{geno}{a genotype matrix obtained with `vcfR`.}

\item{output.file}{a character string indicating the name of the output file.}

\item{allele.sep}{a character indicating which type of delimiter is used to separate the alleles.}

\item{blocksize}{an integer indicating the number of SNPs to be processed in the same chunk.}

\item{console.count}{an integer indicating the number of SNPs by which the progress bar should increment.}
}
\description{
\code{vcf2pcadapt} uses the package \code{vcfR} to extract the genotype information from a vcf file and exports it
under the format required by \code{pcadapt}.
}
\examples{
## see also ?pcadapt for examples

}
\keyword{internal}

