% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vcf_convert}
\alias{vcf_convert}
\title{Convert vcfR genotype matrices}
\usage{
vcf_convert(string_geno, output, allele_sep)
}
\arguments{
\item{string_geno}{a genotype matrix extracted from a VCF file with `vcfR`.}

\item{output}{a character string indicating the name of the output file.}

\item{allele.sep}{a vector of characters indicating what delimiters are used 
to separate alleles.}
}
\description{
\code{vcf_convert} converts outputs of \code{extract.gt} to the format 
\code{pcadapt}.
}
\keyword{internal}
