% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writebed.R
\name{writeBed}
\alias{writeBed}
\title{Write PLINK files}
\usage{
writeBed(x)
}
\arguments{
\item{x}{A [mmapchar][mmapchar-class] object associated 
with a pcadapt or lfmm file.}
}
\value{
The input `bedfile` path.
}
\description{
Function to write bed/bim/fam files from a pcadapt or an lfmm file.
Files shouldn't already exists.
}
