% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedadapt.R
\docType{class}
\name{bedadapt-class}
\alias{bedadapt-class}
\alias{bedadapt_RC}
\alias{bedadapt}
\title{Class bedadapt}
\usage{
bedadapt(path = tempfile(), nIND = NULL, nSNP = NULL, init = NULL)
}
\arguments{
\item{path}{path to the .bed file.}

\item{nIND}{number of individuals.}

\item{nSNP}{number of SNPs.}

\item{init}{not sure what it is made for.}
}
\description{
Class bedadapt

Wrapper constructor for class `bedadapt`.
}

