% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptStat.R
\name{cmpt.stat}
\alias{cmpt.stat}
\title{pcadapt statistics}
\usage{
cmpt.stat(zscores, s.v, K, method, nSNP, maf, min.maf)
}
\arguments{
\item{zscores}{a real-valued matrix obtained from the multiple linear regression of the genotype on the scores.}

\item{s.v}{a vector containing the K larger singular values.}

\item{K}{an integer specifying the number of principal components to retain.}

\item{method}{a character string specifying the method to be used to compute
the p-values. Three statistics are currently available, \code{"mahalanobis"},
\code{"communality"}, and \code{"componentwise"}.}

\item{nSNP}{an integer specifying the number of genetic markers present in the data.}
}
\value{
The returned value is a list containing the test statistics and the associated p-values.
}
\description{
\code{cmpt.stat} returns chi-squared distributed statistics.
}

