% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cmpt_centroids}
\alias{cmpt_centroids}
\title{Ancestral populations centroids}
\usage{
cmpt_centroids(u, lab, pop1, pop2)
}
\arguments{
\item{u}{a matrix of scores.}

\item{lab}{a vector of integers.}

\item{pop1}{an integer.}

\item{pop2}{an integer.}
}
\value{
The returned value is a list.
}
\description{
\code{cmpt_centroids} returns the average scores for each population.
}
