% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{convert.line}
\alias{convert.line}
\title{Convert genotype information obtained with vcfR}
\usage{
convert.line(hap.block, allele.sep = "/")
}
\arguments{
\item{hap.block}{a genotype matrix obtained with `vcfR`.}

\item{allele.sep}{a character indicating which type of delimiter is used to separate the alleles.}
}
\description{
\code{convert.line} converts outputs of \code{extract.gt} to the format \code{pcadapt}.
}
\examples{
## see also ?pcadapt for examples

}
\keyword{internal}

