% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_axis}
\alias{get_axis}
\title{Axis of projection}
\usage{
get_axis(uglob, lab, pop1, pop2)
}
\arguments{
\item{uglob}{a matrix of global scores.}

\item{lab}{a vector of integers.}

\item{pop1}{an integer.}

\item{pop2}{an integer.}
}
\value{
The returned value is a numeric vector.
}
\description{
\code{get_axis} returns the axis onto which projection should be performed.
}
