% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_nb_ind}
\alias{get_nb_ind}
\title{Number of individuals in a specific population}
\usage{
get_nb_ind(lab, anc)
}
\arguments{
\item{lab}{a vector of integers.}

\item{anc}{an integer.}
}
\value{
The returned value is an integer.
}
\description{
\code{get_nb_ind} returns the number of individuals in a specific population.
}
