% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{get_statistics}
\alias{get_statistics}
\title{pcadapt statistics}
\usage{
get_statistics(zscores, method, pass)
}
\arguments{
\item{zscores}{a numeric matrix containing the z-scores.}

\item{method}{a character string specifying the method to be used to compute
the p-values. Two statistics are currently available, \code{"mahalanobis"},
and \code{"componentwise"}.}

\item{pass}{a boolean vector.}
}
\value{
The returned value is a list containing the test statistics and the 
associated p-values.
}
\description{
\code{get_statistics} returns chi-squared distributed statistics.
}
\keyword{internal}
