% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptggPlotting.R
\name{manhattan.plotting}
\alias{manhattan.plotting}
\title{Manhattan Plot}
\usage{
manhattan.plotting(x, K)
}
\arguments{
\item{x}{an object of class "pcadapt" generated with \code{pcadapt} containing the p-values of interest.}
}
\description{
\code{manhattanPlot} displays a Manhattan plot which represents the p-values for each SNP for a particular test statistic.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}

