% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Manhattan Plot}
\usage{
manhattanPlot(x, K)
}
\arguments{
\item{x}{an object of class "pcadapt" generated with \code{pcadapt} containing the p-values of interest.}
}
\description{
\code{manhattanPlot} displays a Manhattan plot which represents the p-values for each SNP for a particular test statistic.
}
\examples{
## see ?pcadapt for examples

}
\keyword{internal}

