\name{pcadapt-package}
\alias{pcadapt-package}
\docType{package}
\title{
	Principal Component Analysis for Outlier Detection.
}
\description{
This package has been developed to provide statistical tools for outlier detection based on Principal Component Analysis.
}
\details{
\tabular{ll}{
Package: \tab pcadapt\cr
Type: \tab Package\cr
Version: \tab 3.0.3\cr
Date: \tab 2016-11-03\cr
License: \tab (>= 2)\cr
}
For an overview of how to use the package, please check the html document provided as a vignette by typing the following command in the \code{R} console:

\code{browseVignettes("pcadapt")} 
}
\author{
Keurcien Luu, Michael G.B. Blum

Maintainer: Keurcien Luu <keurcien.luu@imag.fr>
}
\references{
K Luu, E Bazin and MGB Blum. pcadapt: an R package for performing genome scans for selection based on principal component analysis. biorXiv:10.1101/056135 (2016)
}
\keyword{ package }
\seealso{
http://membres-timc.imag.fr/Michael.Blum/PCAdapt.html
}
\examples{
## see ?pcadapt for examples
}
