\name{pcadapt-package}
\alias{pcadapt-package}
\docType{package}
\title{
	Principal Component Analysis for Outlier Detection.
}
\description{
This package has been developed to provide statistical tools for outlier detection based on Principal Component Analysis.
}
\details{
\tabular{ll}{
Package: \tab pcadapt\cr
Type: \tab Package\cr
Version: \tab 2.1.1\cr
Date: \tab 2016-01-05\cr
License: \tab (>= 2)\cr
}
For an overview of how to use the package, please check the html document provided as a vignette by typing the following command in the \code{R} console:

\code{browseVignette("pcadapt")} 
}
\author{
Keurcien Luu, Michael G.B. Blum

Maintainer: Keurcien Luu <keurcien.luu@imag.fr>
}
\references{
Duforet-Frebourg, N., Luu, K., Laval, G., Bazin, E., & Blum, M. G. B. (2015). Detecting genomic signatures of natural selection with principal component analysis: application to the 1000 Genomes data. arXiv preprint arXiv:1504.04543.
}
\keyword{ package }
\seealso{
http://membres-timc.imag.fr/Michael.Blum/PCAdapt.html
}
\examples{
## see ?pcadapt for examples
}
