% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ped2pcadapt}
\alias{ped2pcadapt}
\title{Convert ped files}
\usage{
ped2pcadapt(input, output)
}
\arguments{
\item{input}{a character string specifying the name of the file to be converted.}

\item{output}{a character string specifying the name of the output file.}
}
\description{
\code{ped2pcadapt} converts \code{ped} files to the format \code{pcadapt}.
}
\keyword{internal}
