% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcadapt.R
\name{pval}
\alias{pval}
\title{Principal Components p-values}
\usage{
pval(x, method, K)
}
\arguments{
\item{x}{an output of \code{pcadapt} containing the test statistics.}

\item{method}{a character string that specifies the method used to compute the p-values.}
}
\description{
\code{pval} computes p-values for each genetic marker, depending on the test statistics distribution.
}
\examples{
data <- read4pcadapt("geno3pops",option="example")
x <- pcadapt(data,K=3)
pval(x,method="mahalanobis",K=3)
}
\keyword{internal}

