% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptRead.R
\name{read.pcadapt}
\alias{read.pcadapt}
\title{File Converter}
\usage{
read.pcadapt(input, type = c("pcadapt", "lfmm", "vcf", "ped", "pool",
  "example"), type.out = c("bed", "matrix"), allele.sep = c("/", "|"),
  pop.sizes, ploidy, local.env, blocksize)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of 
the file to be converted.}

\item{type}{a character string specifying the type of data to be converted 
from.}

\item{type.out}{Either a bed file or a standard R matrix. If the input is
a matrix, then the output is automatically a matrix (so that you don't need
to specify this parameter). If the input is a bed file, then the output is 
also a bed file.}

\item{allele.sep}{a vector of characters indicating what delimiters are used 
in VCF files. By default, only "|" and "/" are recognized. 
So, this argument is only useful for \code{type.in = "vcf"}.}

\item{pop.sizes}{deprecated argument.}

\item{ploidy}{deprecated argument.}

\item{local.env}{deprecated argument.}

\item{blocksize}{deprecated argument.}
}
\description{
\code{read.pcadapt} converts genotype matrices or files to an appropriate
  format readable by \code{pcadapt}. For a file as input, you can return 
  either a matrix or convert it in bed/bim/fam files. For a matrix as input,
  this return a matrix.
}
