% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{read.pcadapt}
\alias{read.pcadapt}
\title{File Converter}
\usage{
read.pcadapt(input, type, local.env, ploidy, pop.sizes = NULL,
  allele.sep = c("/", "|"), blocksize)
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of the file to be
converted.}

\item{type}{a character string specifying the type of data to be converted to the
\code{pcadapt} format. Supported formats are: \code{ped}, \code{vcf}, \code{lfmm}.}

\item{local.env}{deprecated argument.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}

\item{pop.sizes}{a vector specifying the number of individuals to be sampled for each pool.}

\item{allele.sep}{a vector of characters indicating what delimiters are used in VCF files. By default, only
"|" and "/" are recognized.}

\item{blocksize}{deprecated argument.}
}
\description{
\code{read.pcadapt} converts genotype matrices or files to an appropriate
format readable by \code{pcadapt}. You may find the converted file in the
current directory.
}
