% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcadaptUtils.R
\name{read.pcadapt}
\alias{read.pcadapt}
\title{File Converter}
\usage{
read.pcadapt(input.filename, type)
}
\arguments{
\item{input.filename}{a character string specifying the name of the file to be
converted.}

\item{type}{a character string specifying the type of data to be converted to the
\code{pcadapt} format. Supported formats are: \code{ped}, \code{vcf}, \code{lfmm}.}
}
\description{
\code{read.pcadapt} converts \code{.vcf} and \code{.ped} files to an appropriate
type of file readable by \code{pcadapt}. You may find the converted file in the
current directory.
}

