% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/introUtils.R
\name{scan.intro}
\alias{scan.intro}
\title{Introgression}
\usage{
scan.intro(input, K = 2, pop, ancstrl.1, ancstrl.2, admxd, min.maf = 0.05,
  ploidy = 2, window.size = 1000, impute = FALSE, chr.info, map,
  side = "middle")
}
\arguments{
\item{input}{a genotype matrix or a character string specifying the name of 
the file to be imputed.}

\item{K}{a vector of integers specifying the components along which local 
ancestries may vary.}

\item{pop}{a vector of integers or strings specifying which subpopulation the
individuals belong to.}

\item{ancstrl.1}{a string specifying the label of the ancestral population 
genetically closer to the hybrid population.}

\item{ancstrl.2}{a string specifying the label of the ancestral population 
genetically further from the hybrid population.}

\item{admxd}{a string specifying the label of the hybrid population.}

\item{min.maf}{a value between \code{0} and \code{0.45} specifying the 
threshold of minor allele frequencies above which p-values are computed.}

\item{ploidy}{an integer specifying the ploidy of the individuals.}

\item{window.size}{an integer specifying the window size.}

\item{impute}{a logical value indicating whether the input data has to 
imputed.}

\item{chr.info}{a list containing the chromosome information for each marker.}

\item{map}{a numeric vector containing the genetic positions.}
}
\value{
The returned value is a list containing the test statistics and the 
associated p-values.
}
\description{
\code{scan.intro} computes statistics to detect excesses of local ancestry 
based on a PCA approach.
}
