% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updt_local_scores}
\alias{updt_local_scores}
\title{Update local Principal Component Analysis}
\usage{
updt_local_scores(u, geno, V, sigma, beg, end)
}
\arguments{
\item{u}{a score matrix.}

\item{geno}{a genotype matrix.}

\item{V}{a loading matrix.}

\item{sigma}{a vector of singular values.}

\item{beg}{an integer specifying the first marker to be included.}

\item{end}{an integer specifying the first marker to be excluded.}
}
\value{
The returned value is a score matrix.
}
\description{
\code{updt_local_scores} computes the scores using a subset of genetic markers.
}
