% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcfConverter.R
\name{vcf2pcadapt}
\alias{vcf2pcadapt}
\title{vcfR-based converter}
\usage{
vcf2pcadapt(input, output = "tmp.pcadapt", allele.sep = c("/", "|"))
}
\arguments{
\item{input}{a character string specifying the name of the file to be converted.}

\item{output}{a character string indicating the name of the output file.}

\item{allele.sep}{a vector of characters indicating what delimiters are used
to separate alleles.}
}
\description{
\code{vcf2pcadapt} uses the package \code{vcfR} to extract the genotype information 
from a vcf file and exports it under the format required by \code{pcadapt}.
Note that this VCF converter is not maintained anymore. Please use PLINK.
}
\keyword{internal}
